Move ContentPackPersistenceService#loadAllLatest() logic into Graylog
Instead of relying on a (relatively) new version of MongoDB for getting only the
latest revisions of each content pack, the logic was moved into Graylog by loading
all content packs from the database and filter and sort them in memory.

Given the expected low number of content packs per Graylog cluster, this shouldn't
be a problem.

Otherwise, we'd have to bump the minimal system requirements to MongoDB 3.6.x.